Refactor command panel to use same Container/UserInterface separation as other components
