+++ S2VX.Game.Tests/VisualTests/CommandPanelInputBarTests/AddAndEditTests.cs
@@ -3,7 +3,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Testing;
 using osuTK.Input;
-using S2VX.Game.Editor.CommandPanel;
+using S2VX.Game.Editor.UserInterface;
 using S2VX.Game.Story.Command;
 
 namespace S2VX.Game.Tests.VisualTests.CommandPanelInputBarTests {
